refactor: Notification 디렉토리 변경(#304)
