+++ src/main/java/com/soda/member/domain/member/Member.java
@@ -2,7 +2,7 @@ package com.soda.member.domain.member;
 
 import com.soda.common.BaseEntity;
 import com.soda.member.domain.company.Company;
-import com.soda.notification.entity.MemberNotification;
+import com.soda.notification.domain.MemberNotification;
 import com.soda.project.domain.member.MemberProject;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
