+++ src/main/java/com/soda/notification/domain/EmitterService.java
@@ -1,4 +1,4 @@
-package com.soda.notification.service;
+package com.soda.notification.domain;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
