+++ src/main/java/com/soda/notification/infrastructure/NotificationRepository.java
@@ -1,6 +1,6 @@
-package com.soda.notification.repository;
+package com.soda.notification.infrastructure;
 
-import com.soda.notification.entity.Notification;
+import com.soda.notification.domain.Notification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
