+++ src/main/java/com/soda/project/domain/stage/article/comment/Comment.java
@@ -3,8 +3,8 @@ package com.soda.project.domain.stage.article.comment;
 import com.soda.common.BaseEntity;
 import com.soda.member.domain.member.Member;
 import com.soda.project.domain.stage.article.Article;
-import com.soda.notification.event.CommentCreatedEvent;
-import com.soda.notification.event.ReplyCreatedEvent;
+import com.soda.notification.interfaces.event.CommentCreatedEvent;
+import com.soda.notification.interfaces.event.ReplyCreatedEvent;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
